It should also work for different types such as int and float.

+++ Day9/main.cpp
@@ -19,7 +19,7 @@ auto func()
 int main()
 {
     std::array<int, 2> a2;
-    std::array<int, 3> a3;
+    std::array<float, 3> a3;
 
     a2 = func(); // func is gonna deduce the return type's 2 and do plan A
     a3 = func(); // func is gonna deduce the return type's 3 and do plan B
